refactor: 쇼핑몰 결제정보 컴포넌트 분리
